+++ src/main/java/com/noose/todo/domain/note/entity/ScheduleNote.java
@@ -11,8 +11,6 @@ import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-import java.util.List;
-
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Getter
 @DiscriminatorValue("S")
@@ -20,9 +18,6 @@ import java.util.List;
 public class ScheduleNote extends Note {
     @Embedded
     private Period period;
-
-    @Embedded
-    private Todos todos = new Todos();
 
     private ScheduleNote(Long id, Title title, Body body, Period period, Todos todos) {
         super(id, title, body);
@@ -40,28 +35,5 @@ public class ScheduleNote extends Note {
 
     public static ScheduleNote of(String title, String body, Todos todos) {
         return new ScheduleNote(null, new Title(title), new Body(body), new Period(), todos);
-    }
-
-    @Override
-    public List<Todo> todos() {
-        return todos.getValues();
-    }
-
-    public void addTodo(Todo todo) {
-        todos.add(todo);
-        todo.setNote(this);
-    }
-
-    public void addTodos(Todos todos) {
-        this.todos.addAll(todos);
-        todos.setNote(this);
-    }
-
-    public int todoSize() {
-        return todos.size();
-    }
-
-    public void deleteTodo(Todo todo) {
-        todos.delete(todo);
     }
 }
