+++ src/main/resources/application.yaml
@@ -12,4 +12,5 @@ spring:
       hibernate:
         format_sql: true
         show_sql: true
+        default_batch_fetch_size: 100
     open-in-view: false
